+++ Testeeee/src/DAO/DAOProdutosnew.java
@@ -24,7 +24,7 @@ public ArrayList<ModelProdutos> retonarListaProdutosDAO(){
 				modelProdutos = new ModelProdutos();
 				modelProdutos.setIdProduto(this.getResultSet().getInt(1));
 				modelProdutos.setProNome(this.getResultSet().getString(2));
-				modelProdutos.setProValor(this.getResultSet().getDouble(3));
+				modelProdutos.setProValor(this.getResultSet().getFloat(3));
 				modelProdutos.setproEstoque(this.getResultSet().getInt(4));
 				
 				listaModelProduto.add(modelProdutos);
